refactor: rename handleErrorResponse middleware file
