+++ src/pages/playlist/[id].tsx
@@ -14,7 +14,7 @@ import Image from 'next/image';
 import { P, match } from 'ts-pattern';
 import { useAllowedUserInput } from '../../hooks/useAllowedUserInput';
 import { sessionIsActive } from '../../useCases/auth';
-import { handleErrorResponse } from '../../middlewares';
+import { handleErrorResponse } from '../../middlewares/handleErrorResponse';
 import { InternalServerError, UnauthorizedError } from '../../errors';
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
